Move Coupler to cirq_gooogle.ops
- Noticed when adding code that there were circular
imports due to Couplers api codes needing code from devices.
- Moved this to ops directory instead, which avoids the circular
import.
- Note that, although GridQubit is in cirq.devices, LineQubit and
NamedQubit are in cirq.ops, so this also seems to be a good place
to put this class.
