+++ cirq-google/cirq_google/devices/__init__.py
@@ -21,6 +21,4 @@ from cirq_google.devices.google_noise_properties import (
 
 from cirq_google.devices.known_devices import Sycamore, Sycamore23
 
-from cirq_google.devices.coupler import Coupler
-
 from cirq_google.devices.grid_device import GridDevice
