+++ cirq-google/cirq_google/ops/__init__.py
@@ -16,6 +16,8 @@
 
 from cirq_google.ops.calibration_tag import CalibrationTag
 
+from cirq_google.ops.coupler import Coupler
+
 from cirq_google.ops.fsim_gate_family import FSimGateFamily
 
 from cirq_google.ops.fsim_via_model_tag import FSimViaModelTag
